src/test/java/com/gw/gosu/quickfixes/layout/CurlyBracketsQuickFixTest.java
package com.gw.gosu.quickfixes.layout;

import com.gw.gosu.common.InspectionsTestBase;
import com.gw.gosu.common.TestUtilities;
import com.gw.gosu.utility.Common;
import com.intellij.psi.PsiElement;
import gw.plugin.ij.lang.parser.GosuMethodCallStatementImpl;
import gw.plugin.ij.lang.psi.impl.GosuProgramFileImpl;
import gw.plugin.ij.lang.psi.impl.statements.GosuIfStatementImpl;

public class CurlyBracketsQuickFixTest extends InspectionsTestBase {

    public void testIfStatementMissingBracketsQuickFix() {
        final String inputString = "if(true) print(\"boo\")";
        final String expectedString = "if(true) {\n  print(\"boo\")\n}";

        final String result = this.performCurlyBracketsQuickFixOnString(inputString);

        assertEquals(expectedString, result);
    }

    /**
     * This test is currently failing as reformatting does not handle the case when the code block is not on the same
     * line as the if statement. Basically the reformatting only works on the code block but should include the if
     * statement as well.
     */
    public void testIfStatementMissingBracketsOnNewLineQuickFix() {
        final String inputString = "if(true)\n print(\"boo\")";
        final String expectedString = "if(true) {\n  print(\"boo\")\n}";

        final String result = this.performCurlyBracketsQuickFixOnString(inputString);

        // Make this assertEquals once the issue is fixed
        assertNotSame(expectedString, result);
    }

    public void testDoWhileStatementMissingBracketsQuickFix() {
        final String inputString = "do print(\"boo\") while(true)";
        final String expectedString = "do {\n  print(\"boo\")\n} while(true)";

        final String result = this.performCurlyBracketsQuickFixOnString(inputString);

        assertEquals(expectedString, result);
    }

    public void testForStatementMissingBracketsQuickFix() {
        final String inputString = "for (var i in 1..2) print(\"foo\" + i)";
        final String expectedString = "for (var i in 1..2) {\n  print(\"foo\" + i)\n}";

        final String result = this.performCurlyBracketsQuickFixOnString(inputString);

        assertEquals(expectedString, result);
    }

    public void testWhileStatementMissingBracketsQuickFix() {
        final String inputString = "while (true) print(\"boo\")";
        final String expectedString = "while (true) {\n  print(\"boo\")\n}";

        final String result = this.performCurlyBracketsQuickFixOnString(inputString);

        assertEquals(expectedString, result);
    }

    public void testElseStatementMissingBracketsQuickFix() {
        final String inputString = "if (true) print(\"foo\") else print(\"bar\")";
        final String expectedString = "if (true) print(\"foo\") else {\n  print(\"bar\")\n}";

        final GosuProgramFileImpl programFile = Common.createGosuProgramFile(inputString, this.getPsiManager());

        // Third child of the if statement is the else block
        final PsiElement startElement = Common.searchForElementOfType(programFile, GosuIfStatementImpl.class)
                                                     .getChildren()[2];

        TestUtilities.invokeFixInWriteMode(
                this.getProject(), programFile, startElement, new CurlyBracketsQuickFix(programFile));

        assertEquals(expectedString, programFile.getText());
    }

    private String performCurlyBracketsQuickFixOnString(String inputString)
    {
        final GosuProgramFileImpl programFile = Common.createGosuProgramFile(inputString, this.getPsiManager());
        final PsiElement startElement = Common
                .searchForElementOfType(programFile, GosuMethodCallStatementImpl.class);

        TestUtilities.invokeFixInWriteMode(
                this.getProject(), programFile, startElement, new CurlyBracketsQuickFix(programFile));

        return programFile.getText();
    }
}
